// apps/arch-app/src/typechain-types/CourseManager.ts
export declare namespace CourseManager {
  export type CourseStruct = {
    courseId: BigNumberish;
    name: string;
    description: string;
    price: BigNumberish;
    instructor: AddressLike;
    isActive: boolean;
    createdAt: BigNumberish;
    totalStudents: BigNumberish;
    contentHash: string;
    thumbnailHash: string;
  };

  export type CourseStructOutput = [
    courseId: bigint,
    name: string,
    description: string,
    price: bigint,
    instructor: string,
    isActive: boolean,
    createdAt: bigint,
    totalStudents: bigint,
    contentHash: string,
    thumbnailHash: string
  ] & {
    courseId: bigint;
    name: string;
    description: string;
    price: bigint;
    instructor: string;
    isActive: boolean;
    createdAt: bigint;
    totalStudents: bigint;
    contentHash: string;
    thumbnailHash: string;
  };
}